+++ cleo/__init__.py
@@ -4,7 +4,7 @@ import cleo.opto
 import cleo.coords
 import cleo.stimulators
 import cleo.recorders
-import cleo.processing
+import cleo.ioproc
 import cleo.utilities
 import cleo.viz
