+++ cleo/ioproc/__init__.py
@@ -1,0 +1,9 @@
+"""Classes and functions for constructing and configuring an :class:`~cleo.IOProcessor`."""
+from cleo.ioproc.base import (
+    LatencyIOProcessor,
+    ProcessingBlock,
+    RecordOnlyProcessor,
+)
+from cleo.ioproc.controllers import PIController
+from cleo.ioproc.observers import FiringRateEstimator
+from cleo.ioproc.delays import Delay, ConstantDelay, GaussianDelay
