+++ tests/electrodes/test_spiking.py
@@ -4,7 +4,7 @@ import numpy as np
 from brian2 import SpikeGeneratorGroup, ms, mm, Network
 from cleo import CLSimulator
 from cleo.ephys import *
-from cleo.processing import RecordOnlyProcessor
+from cleo.ioproc import RecordOnlyProcessor
 
 
 def _spike_generator_group(z_coords_mm, indices=None, times_ms=None, **kwparams):
